+++ R/aovlist-support.R
@@ -23,7 +23,7 @@
 
 #' @export
 recover_data.aovlist = function(object, ...) {
-    fcall = attr(object, "call")
+    fcall = match.call(aov, attr(object, "call")) # matches even if called via `manova()`
     trms = terms(object)
     # Find the Error terms
     lbls = attr(trms, "term.labels")
@@ -63,35 +63,65 @@ emm_basis.aovlist = function (object, trms, xlev, grid, vcov., ...) {
     # initialize arrays
     nonint = setdiff(names(object), "(Intercept)")
     
-    k = length(xnms)
-    bhat = rep(NA, k) # I'll use NAs to track which slots I've filled
-    V = matrix(0, nrow=k, ncol=k)
-    names(bhat) = xnms
-    dimnames(V) = list(xnms, xnms)
+    k = npar = length(xnms)
+    bhat1 = rep(NA, k) # I'll use NAs in 1st dim of bhat to track which slots I've filled
+    # check for multivariate response
+    m = ifelse (is.matrix(coefm <- object[[1]]$coefficients), ncol(coefm), 1)
+    bhat = rep(NA, k*m)
+    zmm1 = seq_len(m) - 1   # seq 0 : (m-1)
+    
+    
+    #### utility functions...
+    # return indices of as.numeric(x[i, ]) where x has nr rows
+    indx = function(i, nr = npar) 
+        as.numeric(sapply(zmm1, function(j) nr * j + i))
+    
+    # get names or rownames
+    mynames = function(x)
+        if (m == 1) names(x) else rownames(x)
+    
+    V = matrix(0, nrow = k*m, ncol = k*m)
+    names(bhat1) = xnms
+    allxnms = xnms
+    if (m > 1) {
+        ylevs = colnames(coefm)
+        allxnms = as.character(sapply(ylevs, function(.) paste0(xnms, .)))
+    }
+    names(bhat) = allxnms
+    dimnames(V) = list(allxnms, allxnms)
     empty.list = as.list(nonint)
     names(empty.list) = nonint
     Vmats = Vidx = Vdf = empty.list
-    wts = matrix(0, nrow = length(nonint), ncol = k)
-    dimnames(wts) = list(nonint, xnms)
+    wts = matrix(0, nrow = length(nonint), ncol = k*m)
+    dimnames(wts) = list(nonint, allxnms)
     # NOTE: At present, I just do intra-block analysis: wts are all 0 and 1
-    btemp = bhat #++ temp for tracking indexes
+    btemp = bhat1 #++ temp for tracking indexes
     #++Work thru strata in reverse order
     for (nm in rev(nonint)) {
         x = object[[nm]]
+        if (m > 1) class(x) = c("mlm", "lm")   # because vcov.aov is NOT suitable
         bi = coef(x)
-        bi = bi[!is.na(bi)]
-        ii = match(names(bi), xnms)
-        Vidx[[nm]] = use = setdiff(ii, which(!is.na(bhat))) #++ omit elts already filled
+        rn = mynames(bi)
+        nr = length(rn)
+        idx = which(!is.na(bi[seq_along(rn)]))
+        bi = bi[indx(idx, nr)]
+        ii = match(rn[idx], xnms)
+        use = setdiff(ii, which(!is.na(bhat1))) #++ omit elts already filled
         if(length(use) > 0) {
             ii.left = seq_along(ii)[!is.na(match(ii,use))]
-            wts[nm, use] = 1
-            bhat[use] = bi[ii.left]
-            Vi = vcov(x, complete = FALSE)[ii.left, ii.left, drop=FALSE]
+            wts[nm, indx(use)] = 1
+            bhat1[use] = bi[ii.left]
+            allii.left = indx(ii.left, nr)
+            alluse = Vidx[[nm]] = indx(use)
+            bhat[alluse] = bi[allii.left]
+            # following is OK now that we have class(x) = "mlm"
+            Vi = vcov(x, complete = FALSE)[allii.left, allii.left, drop = FALSE]
             Vmats[[nm]] = Vi
-            V[use,use] = Vi
+            V[alluse, alluse] = Vi
         }
         else {
             Vmats[[nm]] = matrix(0, nrow=0, ncol=0)
+            Vidx[[nm]] = integer(0)
         }
         # Any cases with 0 df will have NaN for covariances. I make df = -1 
         # in those cases so I don't divide by 0 later in Satterthwaite calcs
@@ -102,16 +132,24 @@ emm_basis.aovlist = function (object, trms, xlev, grid, vcov., ...) {
     if (!is.null(x)) {
         # The intercept belongs in the 1st error stratum
         # So have to add a row and column to its covariance matrix
-        bhat[1] = x$coefficients[1]
-        wts[1,1] = 1
-        Vidx[[1]] = ii = c(1, Vidx[[1]])
+        idx1 = indx(1)
+        bhat[idx1] = as.numeric(coef(x))
+        wts[1, idx1] = 1
+        x = object[[nonint[1]]] # 1st non-intercept stratum
+        Vidx[[1]] = ii = sort(c(idx1, Vidx[[1]]))
         k = length(ii)
-        vv = matrix(0, nrow=k, ncol=k)
-        if (k > 1) vv[2:k,2:k] = Vmats[[1]]
+### *** FIX THIS ***        
+        vv = matrix(0, nrow = k, ncol = k)
+        i2k = indx(2:(k / m), k / m)
+        if (k > m) vv[i2k, i2k] = Vmats[[1]]
         # Variance of intercept is EMS of this stratum divided by N
         # Here I'm assuming there are no weights
-        N = sum(sapply(object, function(x) length(x$residuals)))
-        V[1,1] = vv[1,1] = sum(object[[2]]$residuals^2) / object[[2]]$df / N
+        N = sum(sapply(object, function(x) length(x$residuals))) / m
+        i1 = indx(1, k / m)
+        if (m > 1)
+            V[idx1, idx1] = vv[i1,i1] = estVar(x) / N
+        else
+            V[1,1] = vv[1,1] = sum(resid(x)^2) / x$df / N
         #dimnames(vv) = list(c(xnms[ii], xnms[ii]))
         Vmats[[1]] = vv
     }
@@ -129,6 +167,10 @@ emm_basis.aovlist = function (object, trms, xlev, grid, vcov., ...) {
     }
     nbasis = estimability::all.estble  # Consider this further?
     misc = list()
+    if (m > 1) {
+        misc$ylevs = list(rep.meas = ylevs)
+        X = kronecker(diag(1, m), X)
+    }
     
     list(X = X, bhat = bhat, nbasis = nbasis, V = V, dffun = dffun, 
          dfargs = dfargs, misc = misc)